+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Fixed
+- [`multipath`] Service name is `multipathd`, aliasing to `multipath-tools` vanishes in Trixie
+
 ## [2.7.0] - 2024-10-10
 
 ### Added